mv: client module -> client.rs
